added security for recette deletion

+++ src/app.js
@@ -22,10 +22,14 @@ app.get('/recettes', cors(), async (req, res) => {
 app.delete('/recettes/:id', cors(), async (req, res) => {
   const id = req.params.id;
   try {
-    const [result] = await db
-      .promise()
-      .query('DELETE FROM recettes WHERE id = ?', [id]);
-    res.status(200).send('deleted');
+    if (id <= 10) {
+      return res.status(401).send('unauthorized');
+    } else {
+      const [result] = await db
+        .promise()
+        .query('DELETE FROM recettes WHERE id = ?', [id]);
+      res.status(200).send('deleted');
+    }
   } catch (error) {
     res.status(500).send('something wrong happened');
   }
